requirements.txt
-i https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cu118
accelerate==0.21.0
aiohttp==3.8.5 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
anyio==3.7.1 ; python_version >= '3.7'
arrow==1.2.3 ; python_version >= '3.6'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==23.1.0 ; python_version >= '3.7'
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
bleach==6.0.0 ; python_version >= '3.7'
blessed==1.20.0 ; python_version >= '2.7'
bokeh==3.2.1
boto3==1.28.19
botocore==1.31.19 ; python_version >= '3.7'
bravado==11.0.3 ; python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_full_version != '3.5.0'
bravado-core==6.1.0 ; python_version >= '3.7'
certifi==2023.7.22 ; python_version >= '3.6'
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
click==8.1.6 ; python_version >= '3.7'
cmake==3.27.0
colorama==0.4.6 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
contourpy==1.1.0 ; python_version >= '3.8'
coolname==2.2.0
cramjam==2.7.0 ; python_version >= '3.7'
datasets==2.14.3
dill==0.3.7
docopt==0.6.2
einops==0.6.1
exceptiongroup==1.1.2 ; python_version < '3.11'
fastparquet==2023.7.0
filelock==3.12.2 ; python_version >= '3.7'
fqdn==1.5.1
frozenlist==1.4.0 ; python_version >= '3.8'
fsspec==2023.6.0 ; python_version >= '3.8'
future==0.18.3 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
gitdb==4.0.10 ; python_version >= '3.7'
gitpython==3.1.32 ; python_version >= '3.7'
gputil==1.4.0
greenlet==2.0.2 ; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0 ; python_version >= '3.7'
h2o-wave==0.26.2
hf-transfer==0.1.3
httpcore==0.17.3 ; python_version >= '3.7'
httpx==0.24.1 ; python_version >= '3.7'
huggingface-hub==0.16.4
idna==3.4 ; python_version >= '3.5'
inquirer==3.1.3 ; python_version >= '3.8'
isoduration==20.11.0
jinja2==3.1.2
jmespath==1.0.1 ; python_version >= '3.7'
joblib==1.3.1 ; python_version >= '3.7'
jsonpointer==2.4
jsonref==1.1.0 ; python_version >= '3.7'
jsonschema==4.18.6 ; python_version >= '3.8'
jsonschema-specifications==2023.7.1 ; python_version >= '3.8'
kaggle==1.5.16
lit==16.0.6
markupsafe==2.1.3 ; python_version >= '3.7'
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4 ; python_version >= '3.7'
multiprocess==0.70.15 ; python_version >= '3.7'
neptune==1.3.1
networkx==3.1 ; python_version >= '3.8'
numpy==1.25.2
oauthlib==3.2.2 ; python_version >= '3.6'
openai==0.27.8
packaging==23.1 ; python_version >= '3.7'
pandas==1.5.3
peft==0.4.0
pillow==10.0.0 ; python_version >= '3.8'
pipreqs==0.4.13 ; python_version >= '3.7'
portalocker==2.7.0 ; python_version >= '3.5'
protobuf==3.20.3
psutil==5.9.5 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyarrow==8.0.0
pyjwt==2.8.0 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-editor==1.0.4
python-slugify==8.0.1 ; python_version >= '3.7'
pytz==2023.3
pyyaml==6.0.1
readchar==4.0.5 ; python_version >= '3.7'
referencing==0.30.0 ; python_version >= '3.8'
regex==2023.6.3 ; python_version >= '3.6'
requests==2.31.0 ; python_version >= '3.7'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rfc3339-validator==0.1.4
rfc3987==1.3.8
rpds-py==0.9.2 ; python_version >= '3.8'
s3transfer==0.6.1 ; python_version >= '3.7'
sacrebleu==2.0.0
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1 ; python_version < '3.13' and python_version >= '3.9'
sentencepiece==0.1.99
setuptools==68.0.0 ; python_version >= '3.7'
simplejson==3.19.1 ; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.0 ; python_version >= '3.6'
sniffio==1.3.0 ; python_version >= '3.7'
soupsieve==2.4.1 ; python_version >= '3.7'
sqlalchemy==1.4.49
sqlitedict==1.7.0
starlette==0.27.0
swagger-spec-validator==3.0.3 ; python_version >= '3.7'
sympy==1.12 ; python_version >= '3.8'
tabulate==0.9.0 ; python_version >= '3.7'
tenacity==8.2.2
text-unidecode==1.3
threadpoolctl==3.2.0 ; python_version >= '3.8'
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tornado==6.3.2 ; python_version >= '3.8'
tqdm==4.65.0
transformers==4.31.0
triton==2.0.0 ; platform_system == 'Linux' and platform_machine == 'x86_64'
typing-extensions==4.7.1 ; python_version >= '3.7'
uri-template==1.3.0
urllib3==1.26.16 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn==0.23.2 ; python_version >= '3.8'
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1 ; python_version >= '3.7'
xxhash==3.3.0 ; python_version >= '3.7'
xyzservices==2023.7.0 ; python_version >= '3.8'
yarg==0.1.9
yarl==1.9.2 ; python_version >= '3.7'
